move main code to updater.js
